refactor : section을 핸들링하는 기능을 LineService 에서 SectionService 로 이동
