+++ src/main/java/subway/section/SectionService.java
@@ -2,6 +2,11 @@ package subway.section;
 
 import org.springframework.stereotype.Service;
 import subway.exception.*;
+import subway.line.LineDto;
+
+import java.util.LinkedList;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class SectionService {
@@ -10,6 +15,37 @@ public class SectionService {
 
     public SectionService(SectionDao sectionDao) {
         this.sectionDao = sectionDao;
+    }
+
+    public void createTerminalSections(LineDto lineDto, Long lineId) {
+        Section upTerminalSection = new Section(lineId, lineDto.getUpStationId(), lineDto.getDistance(), lineDto.getDownStationId());
+        Section downTerminalSection = new Section(lineId, lineDto.getDownStationId(), 0, Section.WRONG_ID);
+        sectionDao.save( upTerminalSection );
+        sectionDao.save( downTerminalSection );
+    }
+
+    public LinkedList<Long> getStationsIdOfLine(long lineId) {
+        List<Section> sections = sectionDao.getSectionsOfLine(lineId)
+                .stream()
+                .collect(Collectors.toList());
+        return sortSections(sections);
+    }
+
+    private LinkedList<Long> sortSections(List<Section> sections) {
+        LinkedList<Long> linkedList = new LinkedList<>();
+        Section currentSection = findSectionByNextId(sections, Section.WRONG_ID);
+        while(linkedList.size() != sections.size()) {
+            linkedList.addFirst(currentSection.getStationId());
+            currentSection = findSectionByNextId(sections, currentSection.getStationId());
+        }
+        return linkedList;
+    }
+
+    private Section findSectionByNextId(List<Section> sections, Long nextId) {
+        return sections.stream()
+                .filter(section -> section.getNextStationId() == nextId)
+                .findAny()
+                .orElse(null);
     }
 
     public boolean insertSection(SectionDto sectionDto, Long lineId) {
